Renaming `constants` module.
